streamlit/requirements.txt
streamlit==1.45.1
python-dotenv==1.0.1
langchain==0.3.25
langchain_openai==0.3.20
langchain-community==0.3.24
ortools==9.11.4210
gTTS==2.5.0
pydub==0.25.1
langchain-google-genai==2.1.3
langchain-core==0.3.64
duckduckgo-search==8.0.2
PyPDF2==3.0.1
plotly==6.1.2
langchain-salesforce==0.2.0
#tavily-python==0.7.5
langchain-tavily==0.2.1
openpyxl==3.1.2
pandas==2.2.1
numpy==1.26.4
psutil==7.0.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10